backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.24.1
pydantic==2.10.6
apscheduler==3.10.4
python-dotenv==1.0.1
vaderSentiment==3.3.2
feedparser==6.0.11
supabase==2.0.3
gotrue==1.3.0
